Remove the Owner class and all references to it
We're going to manager our users through Auth0; we probably aren't even going to store them in the database. Almost every piece of code relating to owners will need to be changed.

To start, this commit gets rid of almost everything to do with owners on the server-side.

There are a few references to owner IDs left in NoteController.java and NoteControllerSpec.java. We'll need to address those when we overhaul the NoteController.

Addresses #21.