Remove the getNoteById endpoint
*This* endpoint, on the other hand, is extraneous. The client was never making requests to this endpoint, and I don't think we have any plans to use it in the future.

Meanwhile, the way that this endpoint requires the note's owner's ID turns out to be a little awkward now that we're getting our owner IDs from Auth0. Generally speaking, in the Auth0 era it's going to be slightly more difficult for the client-side to get information about owners. In particular, we only know owner IDs if Auth0 tells them to us.

I think the information Auth0 gives us about the owners should be sufficient for our use-cases. That being said, we should try not to rely on knowing things about the owner—we're less flexible there than we used to be.

A method of the same name was removed from note.service.ts. (It was never used anywhere.)

Addresses #17, #20