Remove the NoteControllerSetup method in DeathTimer
This method isn't called anywhere; it doesn't really have a reason to exist.

Also, it's unclear whether it's meant for testing or production. As a result, I don't know whether I should pass in a real JwtProcessor or a mock. I'm just going to sidestep that problem by deleting the method entirely.