Add JWTs to all the requisite tests
So, for our new API, the only endpoint on our server that doesn't require a JWT for access is GET api/notes, and that's only if the query parameter status=active is present. (That's the only endpoint that you don't need to be logged-in to see—anyone can look at active notes for any user.)

In the tests for NoteController, we aren't actually testing the GET api/notes?status=active endpoint yet. So, I just added JWTs to all of our tests.

At the moment, for most endpoints, we're only testing the happy-path (the path where there's proper authentication). As a result,  these tests are still all passing, NoteController is just ignoring authentication at the moment.

In the future, we still need to test that these endpoints reject invalid credentials. (We aren't testing that at the moment.)

We also need to test a few cases such as providing credentials for the wrong user, or distinguishing between the different query parameters of GET api/notes.

Addresses #17, #19, #20.