Copy quick-start code; failing due to provider scoping?
Copied over the quick-start code.  All of the providers defaulted to @Injectable(providedIn: 'root'), and were unable to find a provider for router.  Removing this argument (in order to match other functions) caused the tests to be unable to find a provider for what they were testing (i.e. AuthService and InterceptorService).  This feels like a scoping issue, but I'm not sure exactly how to fix it.