Add tests for GET api/notes under a variety of circumstances
Here are eight new tests for the GET api/notes endpoint, bringing the total to nine.

We're testing all possible combinations of the GET api/notes endpoint:

- With a valid JWT, a valid JWT for a different user, and an invalid JWT
- With no `status` parameter, the parameter `status=draft`, and the parameter `status=active`.

we haven't even touched the implementation of NoteController.getNotesByOwner() yet, so some of these tests are failing. (The happy-path tests, where we give proper credentials, happen to already pass—but that's just because getNotesByOwner() isn't checking credentials and is giving out notes to whoever wants them.) 

Addresses #17, #20.