Add tests for GET api/notes without the ownerid query parameter
If you don't specify a particular owner, it's assumed that you want all of the notes in the database.

This is fine if you're only asking for active notes. But if you want draft, deleted, and template notes as well, that's is strictly forbidden under all circumstances, no matter who you're logged in as.