Use assertThrows rather than assertEquals(..., mockRes.getStatus())
When we throw an exception like BadRequestResponse, NoteController doesn't actually change the status of mockRes; Javalin will catch any exceptions and take care of the status codes later.

This means that, for testing, we can't look at mockRes's status code, because it hasn't been set yet; rather, we have to check to see if any exceptions were thrown.

Addresses #17, #20.