Make a NewNote interface
Newly created notes that we haven't sent to the server yet have only a fraction of the fields that a normal note has. Because of this, it makes sense to have a different class for them.

(In particular, having a different class for them makes Karma testing easier 🙃)

(Previously, we were using just one class for both and using a lot of unsafe casting to make up for it.)