Use EventEmitter to tell the DoorBoard page to reload notes.
Previously, the AddNoteComponent had been telling the DoorBoardPageComponent to reload its notes by directly calling the getNotesFromServer method on it. This was giving us strange circular-dependency problems. (It's also a bit brittle.) This commit changes it so that instead, AddNoteComponent emits an Angular event that DoorBoardPageComponent reads. This gets rid of any circular dependencies while keeping everything nice and loosely-coupled 🙂

(This code was tested on the circular-dependency-spike branch first, so it *should* be working. Treat it as unverified, however.)