Alter tests to better reflect functionality
Due to the nature of the test data, it has been difficult to manually attach addDates to the data, and it is not generated automatically when populating the test database.  However, any note added through addNewNote will have the addDate field correctly available, as will any document which was parsed as a note.  As this is only an issue in the context of testing, not actual usage, the relevant tests have been altered--there is still sufficient test coverage to prove that the addDate is correctly created, and maintained through editing.