Make getLoginSub check for null
If the user is not logged in, then their Auth0 profile is null, and you can't read their login sub, which is a field of the Auth0 profile.

Previously, we weren't checking if profile is null, and we were getting null-related errors. (Hundreds of them a second.) Now, we're checking for null, and that problem goes away ;)